moving to workspace, adding olehs suggestions, other nits
