+++ vite-hardhat/package.json
@@ -1,58 +1,20 @@
 {
-  "name": "noir-starter",
-  "version": "1.0.0",
-  "type": "module",
+  "name": "vite-hardhat",
   "description": "A template repository to get started with writing zero knowledge programs with Noir.",
+  "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
   "scripts": {
-    "deploy": "bunx hardhat deploy",
-    "dev": "vite dev"
+    "deploy": "bunx hardhat compile && bunx hardhat deploy",
+    "dev": "bun --filter vite dev"
   },
-  "dependencies": {
-    "@noir-lang/backend_barretenberg": "0.36.0",
-    "@noir-lang/noir_js": "0.36.0",
-    "@noir-lang/noir_wasm": "0.36.0",
-    "@noir-lang/types": "0.36.0",
-    "@nomicfoundation/hardhat-ignition": "^0.15.5",
-    "@nomicfoundation/hardhat-ignition-viem": "^0.15.5",
-    "@tanstack/query-sync-storage-persister": "5.0.5",
-    "@tanstack/react-query": "5.44.0",
-    "@tanstack/react-query-persist-client": "5.0.5",
-    "commander": "^12.1.0",
-    "dotenv": "^16.0.3",
-    "hardhat": "^2.19.2",
-    "hardhat-plugin-noir": "^0.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-toastify": "^9.1.1",
-    "shelljs": "^0.8.5",
-    "tsx": "^4.15.4",
-    "viem": "2.x",
-    "wagmi": "2.10.0"
-  },
+  "type": "module",
   "devDependencies": {
-    "@nomicfoundation/hardhat-chai-matchers": "^2.0.7",
-    "@nomicfoundation/hardhat-ethers": "^3.0.6",
-    "@nomicfoundation/hardhat-network-helpers": "^1.0.11",
-    "@nomicfoundation/hardhat-toolbox-viem": "3.0.0",
-    "@nomicfoundation/hardhat-verify": "^2.0.8",
-    "@nomicfoundation/hardhat-viem": "2.0.2",
-    "@tanstack/react-query-devtools": "5.0.5",
-    "@types/bun": "^1.1.4",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^18.15.5",
-    "@types/react": "^18.0.26",
-    "@types/shelljs": "^0.8.7",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "@wagmi/cli": "^2.1.10",
-    "chai": "^4.2.0",
-    "hardhat-gas-reporter": "^1.0.9",
-    "solidity-coverage": "^0.8.5",
-    "ts-node": "^10.9.1",
-    "typechain": "^8.1.0",
-    "typescript": "^4.9.3",
-    "vite": "^5.0.6"
+    "@types/bun": "^1.1.12",
+    "hardhat": "^2.22.15"
   },
-  "engines": {
-    "node": ">=18.19.0"
+  "peerDependencies": {
+    "typescript": "^5.0.0"
   }
 }
