+++ vite-hardhat/packages/noir/.gitignore
