+++ vite-hardhat/packages/vite/wagmi.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from '@wagmi/cli';
 import { react, hardhat } from '@wagmi/cli/plugins';
-import deployment from './deployment.json';
+import deployment from '../../deployment.json';
 
 export default defineConfig({
   out: 'artifacts/generated.ts',
@@ -8,7 +8,7 @@ export default defineConfig({
     react(),
     hardhat({
       project: '.',
-      artifacts: './artifacts',
+      artifacts: '../artifacts',
       deployments: {
         UltraVerifier: {
           [deployment.networkConfig.id]: deployment.address as `0x${string}`,
